Fixed an issue where a NSUInteger was mistakenly used instead of a float.

+++ CodeSample/CSSTileAsset.h
@@ -9,7 +9,7 @@
 #import "CSSAsset.h"
 #import <GLKit/GLKit.h>
 
-const NSUInteger tileStepSize;
+const float tileStepSize;
 
 @interface CSSTileAsset : CSSAsset
